Rename test suites to avoid underscores
