+++ imagetest/cmd/manager/main.go
@@ -13,18 +13,18 @@ import (
 	"cloud.google.com/go/storage"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/disk"
-	imageboot "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/image_boot"
-	imagevalidation "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/image_validation"
+	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/imageboot"
+	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/imagevalidation"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/metadata"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/network"
-	networkperf "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/network_perf"
+	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/network_perf"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/oslogin"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/security"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/ssh"
-	storageperf "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/storage_perf"
+	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/storageperf"
 	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/windows"
-	windowscontainers "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/windows_containers"
-	windowsimagevalidation "github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/windows_image_validation"
+	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/windowscontainers"
+	"github.com/GoogleCloudPlatform/guest-test-infra/imagetest/test_suites/windowsimagevalidation"
 )
 
 var (
