+++ imagetest/test_suites/imagevalidation/clock_test.go
