+++ imagetest/test_suites/imagevalidation/license_test.go
