+++ imagetest/test_suites/imagevalidation/setup.go
@@ -3,7 +3,7 @@ package imagevalidation
 import "github.com/GoogleCloudPlatform/guest-test-infra/imagetest"
 
 // Name is the name of the test package. It must match the directory name.
-var Name = "image_validation"
+var Name = "imagevalidation"
 
 // TestSetup sets up the test workflow.
 func TestSetup(t *imagetest.TestWorkflow) error {
