+++ imagetest/test_suites/networkperf/startupscripts/netclient_startup.sh
