+++ imagetest/test_suites/networkperf/startupscripts/netserver_startup.sh
