+++ imagetest/test_suites/storageperf/startupscripts/install_fio.ps1
