+++ imagetest/test_suites/windowscontainers/docker_container_test.go
