+++ imagetest/test_suites/windowsimagevalidation/image_validation_test.go
