+++ imagetest/test_suites/windowsimagevalidation/setup.go
@@ -7,7 +7,7 @@ import (
 )
 
 // Name is the name of the test package. It must match the directory name.
-var Name = "windows_image_validation"
+var Name = "windowsimagevalidation"
 
 // TestSetup sets up the test workflow.
 func TestSetup(t *imagetest.TestWorkflow) error {
